Cast strlen result to int.

+++ apps/periodNaive.c
@@ -32,7 +32,7 @@ int searchLongestBorderNaive(char *s, int n) {
 int main () {
     char s[MAX_LINE_SIZE];
     scanf("%[^\n]s", s);
-    int n = strlen(s);
+    int n = (int) strlen(s);
 
     // border
     int r = searchLongestBorderNaive(s, n);
